+++ realm/realm-library/src/objectServer/java/io/realm/SyncConfiguration.java
@@ -73,9 +73,6 @@ import io.realm.rx.RxObservableFactory;
 @Beta
 public class SyncConfiguration extends RealmConfiguration {
 
-    public static final int PORT_REALM = 80;
-    public static final int PORT_REALMS = 443;
-
     // The FAT file system has limitations of length. Also, not all characters are permitted.
     // https://msdn.microsoft.com/en-us/library/aa365247(VS.85).aspx
     public static final int MAX_FULL_PATH_LENGTH = 256;