+++ CHANGELOG.md
@@ -7,6 +7,7 @@
 * Breaking change: Location of Realm files are now placed in `getFilesDir()/<userIdentifier>` instead of `getFilesDir()/`.
   This is done in order to support shared Realms among users, while each user retaining their own local copy.
 * Added support for `PermissionOffer` and `PermissionOfferResponse` to `SyncUser.getManagementRealm()`.
+* Breaking change: `SyncUser.all()` now returns Map instead of List.
 
 ## 2.2.3
 
@@ -15,14 +16,16 @@
 * Fixed native memory leak setting the value of a primary key (#3993).
 * Activated Realm's annotation processor on connectedTest when the project is using kapt (#4008).
 * Fixed bug, preventing Sync client to renew the access token (#4038) (#4039).
+* Fixed "too many open files" issue (#4002).
 
 ### Object Server API Changes (In Beta)
 
 * Exceptions thrown in error handlers are ignored but logged (#3559).
+* Removed unused public constants in `SyncConfiguration` (#4047).
 
 ### Internal
 
-* Updated Realm Sync to 1.0.0-BETA-7.1.
+* Updated Realm Sync to 1.0.0-BETA-7.2.
 * Add a Realm backup when receiving a Sync client reset message from the server.
 
 ## 2.2.2